Split Guest and Non-Guest modes for f-checkout Storybook.
Update version and write change log.
